+++ extensions/project-companion/CHANGELOG.md
@@ -1,3 +1,13 @@
 # Project Companion
 
 ## [Initial Version] - 2024-03-13
+
+## [v1.1.0] - 2024-04-11
+
+### Features
+- Now Show a [confirmAlert](https://developers.raycast.com/api-reference/feedback/alert?q=confirmAlert) when deleting a project
+- Added a `Destructive style` on the delete action
+- Update Readme with "How to use it" description.
+
+### Fixes
+- Fixed a bug where the other link was not displayed in the project details